test: Moved integration tests into separate project
